tinkoff-monitor/requirements.txt
pydantic~=4.3.0
rumps~=0.3.0
tinkoff-investments~=0.2.0b33